Allow use of sparse registries
Registries that use the [sparse registry protocol] show up in
`Cargo.lock` not as `registry+URL`, but as `sparse+URL`. This patch
makes it so that guppy (and hakari in particular in my case) doesn't
error out on lockfiles with dependencies from such registries.

There are two open questions here:

1. Since guppy makes no distinction between the _types_ of registries,
the sparse vs. non-sparse separation is lost once we parse into an
`ExternalSource`, and so when printed back out it'll always be printed
as `registry+URL`.

2. Since the prefix is stripped off, hakari in particular will look for
an entry of the form:

```toml
[registries."reg"]
index = "https://whatever"
```

This is unlike what needs to go in `.cargo/config.toml`, which is:

```toml
[registries."reg"]
index = "sparse+https://whatever"
```

This feels potentially unfortunate.

[sparse registry protocol]: https://doc.rust-lang.org/cargo/reference/registry-index.html#sparse-protocol