Don't print empty diffs
Previously, `cargo hakari generate --diff` would print

    info:
    --- original
    +++ modified

when generate doesn't need anything to change. This is noisy, and also
somewhat confusing as a reader may assume that the presence of the
header suggests _something_ changed and that it was the printing of the
diff that got mixed up somehow. Or that the changes where whitespace
changes or some such.

This change makes it so that if the diff is empty, the command produces
nothing to STDOUT instead. This is also more in line with most other
UNIX utilities.