Remove two unused dependencies (#414)
I noticed that target-spec declares a dependency on unicode-ident but
doesn't seem to actually use it (since
84117fb95932bdfbaff6abb90f2d8e335933ed0c if some cursory git archeology
did not lead me astray).

Then I threw cargo-machete and cargo-shear at the workspace to see if
there's anything else. The only true positive is the *workspace*
dependency on insta not being inherited anywhere. Pretty harmless but
might as well remove it while I'm here.